Fix the issue where rancher-machine does not clean up vSphere VMs that are in failed status.
It is a multi-step process to create a vSphere VM. Previously, we updated the machine in the store only after the whole creation process succeeded, which means that if the VM was instantiated but failed to bootstrap, we did not preserve any updated information about the machine in the store. We did not have the machine ID in the file so we could not find it on the vSphere server side when we removed the machine later. As a result, the VM was left on the vSphere side.

With the fix, we retrieve and set the machine ID as soon as it is available, and save the updated machine to the store even if the creation process fails. This ensures we can properly find and destroy the VM on the vSphere server side.